DCS USB Iot Box/DCS USB Iot Box/CODE/Code_esp_DANH/Box_ESP_Upload_V1.0 .3_84/.vscode/Box_ESP_Upload_V1.0 .3_84/MultiSerial/MultiSerial.ino DCS USB Iot Box/DCS USB Iot Box/CODE/Code_esp_DANH/Box_ESP_Upload_V1.0 .3_84/.vscode/Box_ESP_Upload_V1.0 .3_84/MultiSerial/mymemory.h

+++ DCS USB Iot Box/DCS USB Iot Box/CODE/Code_esp_DANH/Box_ESP_Upload_V1.0 .3_84/.vscode/Box_ESP_Upload_V1.0 .3_84/MultiSerial/mymemory.h
@@ -305,12 +305,12 @@ int write_data(uint32_t addr, uint8_t *reg_, uint32_t size_)
 }
 #endif 
 
-int read_data(uint32_t addr, uint8_t *reg_, uint32_t size_)
+int read_data(uint32_t addr, uint8_t *reg_, uint64_t size_)
 {
    int kq = 0;
-    uint16_t length = size_;    
+    uint64_t length = size_;    
     uint16_t writeBufSz;
-    uint16_t data_offset = 0;
+    uint64_t data_offset = 0;
     uint16_t maxBytes = 256;
     uint32_t addr_new = addr;
     bool ghi;
